Fix a potential problem when either "target.directory" or "cfg.objectsdir" is an empty string (occurs when cfg.targetdir == cfg.location).
When either of them is empty will lead to
<OutDir>\</OutDir> or
<IntDir>\</IntDir>
in the output vcxproj file.

In Windows "\" means the root of the current driver and hence the build outputs (.lib, .dll, .exe) will be incorrectly placed at the root of the current driver but not in the same folder at the vcxproj file.