MASM support for VS projects.
Only gets triggered for projects that contain an asm file.

Respects the `includedirs`, `excludes`, and `defines` directives, as well
as the `FatalWarnings` and `MinimumWarnings` flags. Technically also the
`ExtraWarnings` flag, but the default warning level is already the highest
one so it doesn't really do anything.

Some extra defines are available to asm files:
- `_DEBUG`: Defined for debug builds, undefined for non-debug builds.
- `_WIN32`: Defined for the x32 platform, undefined for other platforms.
- `_WIN64`: Defined for the x64 platform, undefined for other platforms.
- `_EXPORT`: Defined to `EXPORT` for shared libraries, defined to being
  empty for other project types.