When merging two lists, maintain relative order of items.
With some linkers, such as ld, the order in which libraries are
included is important. This change ensures that the relative
order of libraries does not change as multiple "links"
configurations are merged.

The side effect is that all configuration lists will be merged this
way (including, for example, library and include directories). But
order should not matter with these other types of lists.