Add new NoRuntimeChecks flag for Visual Studio projects.
This flag disables the Basic Runtime Checks that are always enabled in  configurations that have optimizations disabled. This is useful to disable when running a "Hybrid" build configuration where you optimize code in projects you currently aren't debugging, but don't optimize the code you are debugging but still want the best performance possible.