Fix undefined behavior when using EditAndContinue with mixed optimized/non-optimized code
Using edit and continue with a mixture of optimized/non-optimized code
isn't officially support and there-fore unstable. Periodically, you'll
encouter crashes (it does work, though). This change accounts for that
instability and allows users to choose whether they want optimized
projects to use edit and continue or not. When they opt out, the
optimized projects will be setup in such a way that they will not expose
compatible code to the edit and continue enabled libs.