Change makefile LINKCMD generation depending on cfg.flags.NoLibGroups
If NoLibGroups is set: $(LIBS) is passed as-is to the linker
Else (i.e. default case): $(LIBS) is wrapped by -Wl,--start-group/-Wl,--end-group.

This forces the linker to iterate over the flags the libraries until no further symbol can be resolved.
For the user, this avoids having to worry about the order and repition of libs passed to the linker.
(Most major IDEs have this behaviour by default).