[vs] Sorting VS project dependencies.
`links` unduplicates links by moving them to the end of the list and
removing the previous entry, but when we generate VS projects we get a
list of all links regardless of configuration. When merging these, if
two unrelated configuration blocks have the same name link, this can
cause project references to move around in the list, causing the
vcxproj and sln to change for no reason.