The for loop in removevalues should also break
To match previous behavior, which was correct, the reverse iteration should also `break` so it doesn't needlessly try to match the entry that shifted into that index. No behavior change, as that entry would have already been checked and not matched.

If the table only has one entry that is removed, this code would be indexing an empty table.